# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Chamados",
        "description": "Documentação da API de Chamados",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/atendentes": {
            "get": {
                "tags": [
                    "Atendentes"
                ],
                "summary": "Lista todos os atendentes",
                "operationId": "33bd1bc7c40ea5213970acc8ea2fb48f",
                "responses": {
                    "200": {
                        "description": "Lista de atendentes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Atendente"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Atendentes"
                ],
                "summary": "Cria um novo atendente",
                "operationId": "4b4ccea6c45147e885ec0eb8d60e6418",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "email"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Nome do Atendente"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "atendente@example.com"
                                    },
                                    "telefone": {
                                        "type": "string",
                                        "example": "123456789"
                                    },
                                    "setor_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "is_gestor": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Atendente criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Atendente"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/atendentes/{id}": {
            "get": {
                "tags": [
                    "Atendentes"
                ],
                "summary": "Mostra um atendente específico",
                "operationId": "3311e63c00585237a55bbca654ee5d4d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do atendente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do atendente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Atendente"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Atendente não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Atendentes"
                ],
                "summary": "Atualiza um atendente existente",
                "operationId": "6f1b3ddc8ccd3ee0127156fc6ea85bc5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do atendente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "email"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Nome do Atendente"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "atendente@example.com"
                                    },
                                    "telefone": {
                                        "type": "string",
                                        "example": "123456789"
                                    },
                                    "setor_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "is_gestor": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Atendente atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Atendente"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Atendente não encontrado"
                    },
                    "400": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Atendentes"
                ],
                "summary": "Deleta um atendente",
                "operationId": "6a16e8795dce57d89d9166d2d323aaaa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do atendente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Atendente deletado com sucesso"
                    },
                    "404": {
                        "description": "Atendente não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Atendente": {
                "required": [
                    "nome",
                    "email",
                    "telefone"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "nome": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "telefone": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}